+++ expressions/constante.hh
@@ -5,11 +5,11 @@ class Constante : public Expression {
 public:
     Constante() = delete;
     Constante(const Constante&) = default;
-    Constante(double valeur);
+    Constante(int valeur);
 
 
-    double calculer(const Contexte& contexte) const override;
+    int calculer(const Contexte& contexte) const override;
 
 private:
-    double _valeur;
+    int _valeur;
 };
