+++ expressions/expression.hh
@@ -5,8 +5,7 @@
 
 class Expression {
 public:
-    virtual double calculer(const Contexte & contexte) const = 0;
+    virtual int calculer(const Contexte & contexte) const = 0;
 };
 
 using ExpressionPtr = std::shared_ptr<Expression>;
-
