+++ expressions/variable.cc
@@ -2,6 +2,6 @@
 
 Variable::Variable(const std::string& nom) : _nom(nom) {}
 
-double Variable::calculer(const Contexte & contexte) const {
+int Variable::calculer(const Contexte & contexte) const {
     return contexte[_nom];
 }
